Update tests to entity-based modules.
